fix delete

+++ task-service/src/main/java/com/stacklog/task_service/model/service/TaskService.java
@@ -147,7 +147,9 @@ public class TaskService implements IService<Task> {
             e.getAssigns().stream().forEach(a -> taskAssignService.save(a, token));
         }
         if (e.getReviews() != null && !e.getReviews().isEmpty()) {
-            reviewService.deleteReviews(e.getTaskId(), e.getReviews());
+            if (!isCreate) {
+                reviewService.deleteReviews(e.getTaskId(), e.getReviews());
+            }
             e.getReviews().stream().forEach(r -> reviewService.save(r, token));
         }
         if (e.getCheckLists() != null && !e.getCheckLists().isEmpty()) {
